Split Record provider into controller/provider
